main/requirements.txt

-f https://download.pytorch.org/whl/torch_stable.html
pytesseract
easyocr
image_to_numpy
mmdet==2.3.0
torch==1.5.1+cpu 
torchvision==0.6.1+cpu
opencv_python
image_to_numpy
networkx
-f https://openmmlab.oss-accelerate.aliyuncs.com/mmcv/dist/index.html
mmcv-full==1.0.5+torch1.5.0+cpu
